refactor: moved commit into repo class
